api/requirements.txt
fastapi>=0.115
uvicorn>=0.30
httpx>=0.27
beautifulsoup4>=4.12
readability-lxml>=0.8
tenacity>=9.0
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
langchain>=0.2
langchain-openai>=0.2
langgraph>=0.2
playwright>=1.48
playwright-stealth>=1.0.6

